Bugfix and rewriting in OOP style adding tests by Niyaz Nigmatullin

+++ src/Expression3.java
@@ -3,5 +3,5 @@
  * @version $$Id$$
  */
 public interface Expression3 {
-    double evaluate(double x, double y, double z);
+    int evaluate(int x, int y, int z);
 }
