+++ src/Multiply.java
@@ -5,7 +5,7 @@ public class Multiply extends AbstractBinaryExpression {
     }
 
     @Override
-    public double evaluate(double x, double y, double z) {
+    public int evaluate(int x, int y, int z) {
         return first.evaluate(x, y, z) * second.evaluate(x, y, z);
     }
 }
