+++ src/Negative.java
@@ -8,7 +8,7 @@ public class Negative extends AbstractUnaryExpression {
     }
 
     @Override
-    public double evaluate(double x, double y, double z) {
+    public int evaluate(int x, int y, int z) {
         return - first.evaluate(x, y, z);
     }
 }
